# pnmc-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PNMC Migration API
  version: 1.0.0
  summary: Contrato API inicial para migracion incremental PNMC
  description: |
    Contrato base para desacoplar integraciones sensibles del frontend actual,
    habilitando una migracion paralela sin tocar la UI existente.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-dev.pnmc.local
    description: Dev (referencia)
...
paths:
...
  /api/v1/festivals:
    get:
      tags: [festivals]
      summary: Listar festivales desde SQL
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Festivales paginados
  /api/v1/music-schools:
    get:
      tags: [music-schools]
      summary: Listar escuelas de musica desde SQL
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Escuelas paginadas
  /api/v1/music-markets:
    get:
      tags: [music-markets]
      summary: Listar mercados musicales desde SQL
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Mercados paginados
  /api/v1/organizations:
    get:
      tags: [organizations]
      summary: Listar organizaciones desde SQL
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Organizaciones paginadas
  /api/v1/spaces-infrastructure:
    get:
      tags: [spaces-infrastructure]
      summary: Listar espacios e infraestructura desde SQL
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Espacios paginados
  /api/v1/divipola/locations:
    get:
      tags: [divipola]
      summary: Listar ubicaciones DIVIPOLA
      parameters:
        - in: query
          name: departmentCode
          schema: { type: string }
        - in: query
          name: municipalityCode
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Ubicaciones paginadas
...
  /api/v1/process-entity-relations:
    get:
      tags: [process-entity-relations]
      summary: Listar relaciones proceso-entidad
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Relaciones paginadas
  /api/v1/process-relations:
    get:
      tags: [process-relations]
      summary: Listar relaciones entre procesos
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Relaciones paginadas
...
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
...